requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.3.4
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cliff==2.9.1
cmd2==0.7.7
contextlib2==0.5.5
cryptography==2.0
debtcollector==1.17.0
decorator==4.1.2
deprecation==1.0.1
env==0.1.0
futurist==1.4.0
gnocchiclient==4.0.0
idna==2.6
iso8601==0.1.12
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==3.2.0
monotonic==1.3
msgpack-python==0.4.8
murano-pkg-check==0.3.0
netaddr==0.7.19
netifaces==0.10.6
openstacksdk==0.9.17
os-client-config==1.28.0
osc-lib==1.7.0
oslo.config==4.11.0
oslo.context==2.17.0
oslo.i18n==3.17.0
oslo.log==3.30.0
oslo.serialization==2.20.0
oslo.utils==3.28.0
pbr==3.1.1
ply==3.10
positional==1.1.1
prettytable==0.7.2
pycparser==2.18
pyinotify==0.9.6
pyOpenSSL==17.2.0
pyparsing==2.2.0
pyperclip==1.5.27
python-ceilometerclient==2.9.0
python-cinderclient==3.0.0
python-dateutil==2.6.1
python-designateclient==2.7.0
python-glanceclient==2.7.0
python-heatclient==1.10.0
python-keystoneclient==3.12.0
python-magnumclient==2.6.0
python-mistralclient==3.1.0
python-muranoclient==0.13.0
python-neutronclient==6.3.0
python-novaclient==9.0.1
python-openstackclient==3.11.0
python-saharaclient==1.2.0
python-swiftclient==3.3.0
python-troveclient==2.10.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requestsexceptions==1.3.0
rfc3986==1.1.0
semantic-version==2.6.0
simplejson==3.11.1
six==1.11.0
stevedore==1.27.1
ujson==1.35
urllib3==1.22
warlock==1.2.0
wrapt==1.10.10
yaql==1.1.3
